Exercises 4.15.-4.23
plus added Exercises [4.5, 4.6,4.7]

+++ controllers/blogs.js
@@ -1,36 +1,66 @@
-const blogsRoutes = require('express').Router();
-const Blog = require('../models/blog.js');
+const blogsRouter = require('express').Router();
+const { userExtractor } = require('../utils/middleware');
 
-blogsRoutes.get('/', async (request, response) => {
-  const blogs = await Blog.find({});
+const Blog = require('../models/blog.js');
+const User = require('../models/user.js');
+
+blogsRouter.get('/', async (request, response) => {
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 });
 
   response.json(blogs);
 });
 
-blogsRoutes.get('/:id', async (request, response) => {
+blogsRouter.get('/:id', async (request, response) => {
   const { id } = request.params;
 
   const selectedBlog = await Blog.findById(id);
   response.json(selectedBlog);
 });
 
-blogsRoutes.post('/', async (request, response) => {
-  const { body } = request;
+blogsRouter.post('/', userExtractor, async (request, response) => {
+  const { title, author, url, likes } = request.body;
+
+  if (!request.user.id) {
+    return response.status(401).json({ error: 'token missing or invalid' });
+  }
+  const user = await User.findById(request.user.id);
+
   const blog = new Blog({
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes || 0
+    title,
+    author,
+    url,
+    likes: likes || 0,
+    user: user._id
   });
 
   const savedBlog = await blog.save();
+
+  user.blogs = user.blogs.concat(savedBlog._id);
+  await user.save();
+
   response.status(201).json(savedBlog);
 });
 
-blogsRoutes.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
   const { id } = request.params;
-  await Blog.findByIdAndRemove(id);
-  response.status(204).end();
+
+  if (!request.user.id) {
+    return response.status(401).json({ error: 'token missing or invalid' });
+  }
+  const blog = await Blog.findById(id);
+
+  if (!blog) {
+    return response
+      .status(400)
+      .json({ error: 'blog does not exist or has been deleted' });
+  }
+
+  if (blog.user.toString() === request.user.id.toString()) {
+    await Blog.findByIdAndRemove(id);
+    response.status(204).end();
+  } else {
+    response.status(401).end();
+  }
 });
 
 const options = {
@@ -39,7 +69,7 @@ const options = {
   context: 'query'
 };
 
-blogsRoutes.put('/:id', async (request, response) => {
+blogsRouter.put('/:id', async (request, response) => {
   const { title, author, url, likes } = request.body;
   const { id } = request.params;
 
@@ -49,4 +79,4 @@ blogsRoutes.put('/:id', async (request, response) => {
   response.json(updatedPerson);
 });
 
-module.exports = blogsRoutes;
+module.exports = blogsRouter;
